Fix delete command to work with threads.

+++ src/main/java/com/bp3x/raidbot/commands/lfg/DeleteCommand.java
@@ -6,7 +6,9 @@ import com.bp3x.raidbot.util.RaidBotRuntimeException;
 import com.jagrosh.jdautilities.command.Command;
 import com.jagrosh.jdautilities.command.CommandEvent;
 import net.dv8tion.jda.api.entities.Message;
+import net.dv8tion.jda.api.entities.MessageActivity;
 import net.dv8tion.jda.api.entities.channel.concrete.ThreadChannel;
+import net.dv8tion.jda.api.events.message.MessageUpdateEvent;
 import net.dv8tion.jda.api.exceptions.ErrorResponseException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -58,13 +60,22 @@ public class DeleteCommand extends Command {
            MessageUtils.sendAutoDeletedMessage("Invalid arguments. Use `!delete <eventID>`", 300, commandEvent);
         }
         if (eventToDelete != null && messageToDelete != null) {
-            ThreadChannel thread = messageToDelete.getStartedThread();
-            if (thread != null)
-                thread.delete().queue();
-            else
-                log.error("Thread was null!");
-            messageToDelete.delete().queue();
-            MessageUtils.sendAutoDeletedMessage("Deleted event with ID: " + eventToDelete.getEventId() + ". Please manually delete the thread. We're working on a fix to automatically delete it.", 300, commandEvent);
+            Message latestMessageInstance = commandEvent.getChannel().retrieveMessageById(messageToDelete.getId()).complete();
+            if (latestMessageInstance != null) {
+                ThreadChannel thread = latestMessageInstance.getStartedThread();
+                if (thread != null) {
+                    thread.delete().complete();
+                    messageToDelete.delete().queue();
+                    MessageUtils.sendAutoDeletedMessage("Deleted event with ID: " + eventToDelete.getEventId() + " and its corresponding thread.", 300, commandEvent);
+                }
+                else {
+                    messageToDelete.delete().queue();
+                    MessageUtils.sendAutoDeletedMessage("Deleted event with ID: " + eventToDelete.getEventId() + ". Please delete the thread manually, there was a problem deleting it.", 300, commandEvent);
+                }
+            }
+            else {
+                MessageUtils.sendAutoDeletedMessage("There was a problem locating the message to delete the thread. Contact a mod.", 300, commandEvent);
+            }
         }
     }
 }
